Wait for container to finish output logging
Previously the behavior seemed to be returning logs at the given point
in time we captured them. However, if the container took more than a few
seconds to run we wouldn't get the rest of the logs. Now we wait for the
container to finish and then return those logs.

- Also return stderr to user instead of only stdout