Stream log response back
This was done for 2 reasons:
1. This provides output from the container faster rather than having to
wait for the entire thing to finish.
2. The high priority reason this was persued was that it was causing
nework timeouts for long running containers. While this won't fully
prevent that it will help. Also we don't necessarily want long running
containers but we'd like to have more control over that rather than some
external network setting. I was unable to determine what this network
setting was or how to control it.