Don't stream docker logs back in response
Streaming logs prevents us from sending a success or failure status code
back in the response. The status code header is sent and then all of the
log output. This means that if there was an issue with the build and it
exited with a non-zero exit code there's no way for us to communicate
this with the status code.

Additionally, it isn't strictly necessary for logs to be streamed back
in the response. Rather, it's just a nice to have common feature of a CI
server. However, I think it's a bit of an anti pattern to have to sit
and watch logs on your CI server. See issue #2 for more information

See the following commit for more information on why we decided to
return a streamed response originally:

37b9b059726dc2792a749934a3dfd957596e0ef1