Allow for privisioning a machine to install and start the app
Installs docker (the only real needed dependency) and then runs the
docker-ci-prototype service on port 8080. If the app is already running
then it stops the current one first.

Due to the fact that we stop the currently running one there will be the
possibility that we will fail out if other builds are running. For now
this is a fairly insignificant issue. However, it is not robust.