Use application plugin instead of capsule
Capsule seems unmaintained and was causing issues. I've had strange
issues with it in the past as well. The application plugin is more
straightforward and I think makes more sense to use in our case since we
are packaging everything up in a container and don't necessarily need a
single fat jar.