Run build on alternate build server
When the server builds and deploys itself it kills the existing server
and starts up a new one. This means that the client waiting for a reply
from the build server will have it's connection cut off. In order to
prevent this from happening when building and deploying this project we
will do it on an alternative server which is managed seprately.