Return 400 when the container exits with non-zero status code
We were ignoring the exit code of the container. We'd like the requester
to more easily know when their job failed.

Note that we need to read the logs fully before getting the exit code
because this causes the code to block until the job has finished. This
may not be the most elegant and clear solution.

See issue #2 for more information.