Use wget instead of curl to call the build server
The main reasoning behind this is so we can fail the build when the
build failed. The build server is now returning 400 status codes when
the build failed. curl was unable to exit with a non-zero status code
when the server returned a failure status code response. curl's --fail
flag sounded like it wasn't quite meant for this plus it doesn't print
out the response from the server so it didn't work in this use case.