Revert to using curl to call build server
wget was having too many issues even though it seemed to more natively
support failing when the server returned an error status code.

Instead we now have to use a slightly more complicated build script but
it's satisfactory for now.

The build script writes out the status code as the last line of output
and writes everything to a file. Then we check to see if the last line
was 200. Otherwise, we exit with a code of 1