Generate version prop file on build
This property file will be included in the jar for the application. It
will be accessible to the application to read as a property file. The
property file will include the git version of the build.

The purpose of this file will be to check the version of the server
easily. It will be available by a healthcheck endpoint.