Mount secrets volume if available
With the introduction of the `/secret` endpoint users can now create
secrets to be made available for a specified container. This change
mounts the secrets volume if one has been created.

Currently users can pass in any input, including secrets, to be run with
a container. However, there are situations where a user may want to run
a container which requires permissions that this user doesn't have. This
feature allows the user who does have these permissions to attach them
to a specified container so that anyone is able to run it.