Import route53 zone if it already exists
The zone is shared between the alternate server setup
(alt.simple-ci.com) and the main server (api.simple-ci.com). This means
that if one is created before the other there will be a conflict in
terraform because it doesn't know about the other. This somewhat
gracefully handles this situation by importing the zone if it's already
there.