Limit to 5 concurrent jobs
If too many jobs are running on the server it may become overloaded.
This will help prevent that.

It may also be a good idea to add some kind of queue so that jobs aren't
just rejected. It would be a good idea to have a relatively short
timeout so jobs aren't sitting in the queue for too long.

Note that this doesn't necessarily mean that only 5 docker containers
can be running on the server. Jobs are still able to spin up their own
containers.