Rate limit requests per second
This rate limits to 5 requests per second. Since the service is also
currently configured to allow for only 5 jobs to run concurrently, this
doesn't have too much of an effect, but it will at least have the pieces
in place if these configurations change in the future. These values will
eventually be pulled from the global application configuration.

Note that this combines the functionality of the concurrent job limit
with the rate limit. They needed to be combined so the counters were
incremented atomically. Otherwise one limiter may succeed and increment
it's job request counter while the other limiter fails and doesn't. This
would cause one to count a job which actually wasn't ever run.