Install sysbox container runtime
The Ubuntu cloud image that we are using doesn't have support for
`shiftfs`. This means that when we install the sysbox runtime it
configures docker to be in userns-remap mode. So when we run our
application container we need to specify `--userns=host`. Otherwise,
once we're in the container we won't be able to connect to the mounted
`docker.sock` since we'll have limitations compared to a normal
container. Read more here:

https://github.com/nestybox/sysbox/blob/master/docs/user-guide/install.md#docker-userns-remap

Once we can get `shiftfs` enabled on the cloud server we can remove the
`userns=host` flag.