Wait for the container to finish before accessing it's logs
Once we started using the sysbox runtime, logs starting acting strange.
I didn't do enough investigation but there may have always been an issue
with how we were getting logs.

In any case, this seems to fix the issue that was causing us to lose
logs. It my introduce issues later if we want to bring back the
streaming feature but that will be dealt with when it comes.