Build the build image on new alt conjob server
The build was failing because the main conjob server needed to be
updated. However, since it builds itself there was no way to get a
change to it.

To solve this general problem we should just always build everything
related to conjob on the alt server. It can be updated out of band if
there's a fix needed.

See this commit for more details on why we decided to use the alt server
originally:

ebb7af8f1585f77add1d866c633dfd4fc318d70e