Require tag to be specified
It seems that if a tag isn't specified then it tries to pull down every
single tag. This is definitely not desireable.

This is a temporary fix until we make the change to default to "latest"
tag if it isn't specified by the user.

Perhaps this validation logic can still be reused lower down in the code
just to make sure this doesn't happen before running the image.