Pull secret volume intermediary image if absent
Rather than pulling the secret volume intermediary image every time on
startup, we'll now just pull it if it's missing for every request.

It might also be a good idea to add this pull if absent logic on startup
as well. That way even from the first request to the secret endpoint it
will always be fast. This may negatively affect performance tests in the
future.