Increase secret perf test mean response time allowed
For some reason the response time goes up quite dramatically for each
new concurrent request you add. It doesn't seem to be the case locally
so I don't think it's a network issue. I believe it just has to do with
the lower performance of the AWS EC2 instance tier that we use.