Update job model to be a job run
A job was too broad of a concept as we were using it. A job is basically
the docker image and a run is the container. So what we were really
referring to is the job run rather than the job.

This change updates the job model and all the cascading changes
resulting from it.