Run configure and run docker image all at once
Rather than building the 3rd-party's docker configuration and then run
it, we now create a configuration which can be used to decide how to
configure and run it all at once. This allows us to use our own domain
models as well as create an interface around the docker adapter which
doesn't take or return any third party values and will prevent us from
needing to mock what we don't own.