Add mockito for mocking in unit tests
Although we want to strongly avoid mocking, there are some cases where
we might need to. I'm planning on using them for testing integrations
that may be difficult to test such as that with docker. Although ideally
this would just be done with integratin tests. In any case, the more
tests the better!