Throw exception if JobRunOutcome has an invalid exit code
Before if the exit code was -2 or lower it would result in a SUCCESS
outcome. Although unlikely, this is not strictly correct.

This change adds a new type of test as well: Property based testing.
Instead of specifying specific values for some tests, we are having the
property based testing framework specify a (mostly random) range of
values. This will help cover edge cases.