Simplify alternate bootstrap creation process
The previous approach was still mounting the pwd. This meant that any
changes you had locally would have been applied to the bootstrap server.
We probably would just want to use the last pushed version of the code.

It was also necessary to temporarily edit ./test.sh by hand to have it
use the alt-env rather than the normal one.