Allow config models to be instantiated with values
This retains the no arg constructor and adds the all arg constructor. It
seems like a good idea to allow the configs to be instanatiated with all
values. However, this was specifically done for testing purposes so that
the property based testing framework will be able to create them
automatically.