Remove unused or duplicate methods in the job service
There were duplicate methods that did the same thing but with different
names and some unused methods which would have no future use. There were
also some methods which defaulted empty input to an empty string which
I'm not exactly sure is what we want. Right now we consider null input
to be the deciding factor of when there is input.