Integration test different application configurations
These tests spin up the application in memory and make actual requests
against them. RREST-assured has been added as a dependency in order to
make these requests.

Application now exposes the environment it's initialized with. This is
necessary to shut down the service after each test. Particularly
important so that it frees up ports.