Mock MDC calls to make tests more stable
I believe the MDC is scoped to the entire JVM and persists state until
all tests are finished. Whatever the case, these tests were very flakey
due to the fact that we were using the real MDC which is better off
mocked anyways.

This takes advantage of static mocks which are a little complicated, but
unfortunately necessary in this case.