Revert "Attempt to fix unit test JRE crash"
This reverts commit cf71edfb3c4753cbfab171805062864fbf438fce.

This is being reverted for now since it was only a test to see if it
would fix the JRE crash problem with unit tests. I'm going to try a few
other approaches that wouldn't require us to update java versions since
that's not something I'd like to do right now.