TEMP COMMIT to help troubleshoot JRE error
The config task test is always skipped and the other one is always run
before it so maybe around here is where the problem is. Trying to
exclude them and see what happens. Also remove some other stuff to make
the build be quicker so we can iterate on this and get feedback quicker.