Copy cloned repo to another location before starting the build inside it
This is mainly done to prevent issues with local testing. When local
testing, we mount our current workspace on the build container. Whenever
files would change, such as config files we had to modify, it would
cause them to change in the users workspace which would clutter it up.

Doing things this way also guarantees a clean directory after the build
has finished.