Add HTTPS support to service
The service is started with another environment variable, the keystore
password, which it needs to use the keystore which is copied into the
container before it's started.

Also update tests so they use the HTTPS endpoint (not yet for the admin
endpoint) and don't require strict cert validation in our acceptance
tests since the application is using a cert for HTTPS which is self
signed.