Use HTTPS for prod endpoint
We use use a provider to do the necessary work with Let's Encrypt to
create a key and have it signed by them. Terraform then exports the PKCS
12 contents which ultimately are used by conjob for handling HTTPS.

Note that we're still using the staging endpoint to create the
certificate so it won't be valid yet. We're also still using a stand-in
hardcoded password. This will be updated at a later point after it's
been deployed and validated manually.

Also note that there are some unique terraform changes because it's
version was updated which required some syntax changes.