Use a random password for the prod cert and keystore passwords
We weren't able to use a different password for the testing and
production environments. When the build is run, the same input password
is used in both the testing environment and the production environment.
I think this actually highlights a problem with out setup because all of
the credentials are shared between prod and testing which isn't a good
idea.

In any case, until that problem is solved we now generate a random
password for the production cert and keystore since it isn't important
for us to know what it is nor pass it in at the start of the build.

Instead of getting this value from the build input, we get it from
terraform. Terraform generates a random password from prod, and like the
cert, we hardcode the output value to the password that goes with the
hardcoded cert. This solution seems elegant too because the hardcoded
cert and password live together.