Sync most recent certificate tf changes to alt-env
Note that right now even though we're pointed at the prod instance of
lets encrypt we're using a staging certificate. I believe this is
because we previously requested a certificate with staging and since
terraform doesn't do anything if the certificate doesn't need to be
renewed then it won't try to read the config and fetch from the new URL
we've configured which is the prod lets encrypt endpoint.

I believe this can be fixed by forcing a refresh of the cert, but right
now we're being rate limited because we've tried this 5 other times in
the past 7 days. We can either wait ~60 days for the cert to
automatically renew, or we can try again in 7 days at most to force a
refresh and hopefully that should work.

In any case, it isn't essential that we have a valid cert for the alt
environment.