Add .terraform.lock.hcl files
These files are now created on the version of terraform we use. It's
considered best practice to check in these files as it would be for lock
files in a package manager.

I have a few doubts and we should probably be specifying provider
versions explicitly (which we may not be doing), but I'll check them in
for now and perhaps change this if there are problems later on.