Cache docker containers and images across job runs
Every time a job runs, it runs completely independently of other runs.
While this can be good, it also has significant impacts. The duration of
the run and the disk space taken up will be much longer because there
will be no docker cache.

This change mounts a volume on /var/lib/docker on the job container so
that there will be a cache between runs. For isolation purposes, each
job will have it's own cache volume. This is to prevent possible
unintended behavior and also for security purposes so different jobs
don't have access to each others cache which could have sensitive
information on them such as secrets.

The result of this change should be jobs running quicker and taking up
less space.