Allow removing volumes by name via an admin task
This adds an admin task endpoint which can be called with a volume name
to remove that volume.

Due to a limitation with the underlying lib the volumes can't be force
removed so the associated containers also need to be removed.

This change is to assist with testing the new docker cache
functionality.