Git ignore local-config.yml
The new flow will to be to use this config when working locally rather
than the default. This is typical best practice for working with configs
locally so that you can edit them without potentially checking them in
or having them clutter your index.

This is particularly useful now that we have https configuration that
doesn't work locally.