Use different config file for production
Only production needs https configuration and if it isn't specified the
service won't start.

This prevented the service from starting at all without specifying https
configuration and copying in a key file.

Perhaps it may be a better idea to just build in a key file and password
that can be used for testing purposes. If there are issues with this
implementation that may be considered.