Keep files dir after deployment
The config file that's copied into the container now lives here in
source control. So we need to keep this dir.

The problem was that during the test deployment this dir was deleted and
upon doing the prod deployment this file was no longer there.

We still need to delete the keystore password or for the prod deployment
a new keystore won't be created from keytool.