Stop the alt server when the spot instance is interrupted
Previously our spot instances were terminated. This changes it so the
spot instance is just shut down. It will be automatically restarted when
there is capacity again.

This is a precurser to the work to start conjob automatically on server
boot.

Note that the server may/will have a different IP when it starts back
up and this will need to be resolved so DNS will still work.