Use static IP for alt-env
The alt-env will periodically stop and restart due to the nature of spot
instances. However, we want the service to start back up and be
available once it does start up.

One problem is that once the instance starts back up, it may have a
different IP. The DNS is configured statically to point to an IP and we
don't want to update it out of band with terraform.