Decrease alt-env spot price
Now that the server reboots smoothly after being shut down we can reduce
the price since it isn't necessary to be available constantly. Let's try
to see how far down we can get this before we either won't be able to
get an instance or where it's shutting down often enough where it
becomes a problem.