Temporarily use t3.large instance
We have a job that's hanging and it might be performance related. This
instance is too expensive to use long term, but the spot instance price
would be acceptable. So if this fixes the problem maybe we'll start
using spot instances in prod which is feasible since they can auto start
on reboot now.