Use a spot instance for prod
The t3.small instance seems to be a little shy of resources we need. Now
that we have a setup that allows for instances to shut down and come up
without manual intervention and spot instances auto recover, it's safe
to move to a spot instance. It'll cost less/comparible for more power.