Use a t3.medium for alt-env
Ideally since we rarely use it we'd be able to stick with the small.
However, it's constantly freezing up cuz it's getting overloaded so
we'll have to use a larger instance for now.

Perhaps if/once we add the ability to throttle containers by controlling
their available resources this won't be a problem.