Fix issue with backing docker lib and docker 25+
The backing docker lib we use hasn't been maintained for a while and
needs to be replaced. We've known this for a while, but it seems like
with version 25 there are breaking changes.

This pins our docker version to the latest of 24. We should have been
pinning versions anyways to avoid surprise breaking changes like this
anyways.