Update alt-env deployment for self-hosted variant
This server is often unused and expensive so I'm now self hosting in.

We're no longer using terraform for the alt env. One thing to note is
that the actual setup of the server is still done manually out of band.
This mainly involves setting up the server and putting the required ssh
key on it. Everything else is still handled by the alt-env.sh script
after the machine exists.

- The host is ubuntu 24.04 so we're also bumping the docker build image.
  This was necessary in particular due to an ansible version mismatch.
- We generate our own key for https. We don't care that it isn't valid
  since we're the only ones using it.
- Remove all alt-env terraform code. Eventually we should make something
  that automatically creates the machine.