Fix test so two container IDs are never the same
The code this is testing works by removing containers associated with
volumes after the volume removal failed due to the association. The
property based testing framework was injecting two container ids that
were the same. However, in practice, docker would never give two
container ids that are the same in the error message. So this is a safe
assumption to make in the test.