Update dropwizard: 2.1.12 -> v3
This mostly involved updates to security/auth related classes.

It also required a fix to include runtime classpaths to unit tests. This
was due to a logback change, but we probably should have always been
including them this way.

It also required us to ignore some HTTPS checking in the non-prod
environment due to a upgrade to jetty in dropwizard.