Upgrade build image to Java 21
This mainly involved a large upgrade to gatling. Gatling gradle changes
were necessary for logging and to make it so all tests are run by
default, otherwise it interactively prompts you which tests to run. The
variable syntax also changed from using a $ to #.

Finally, there was a small issue with the tests hitting limits when they
shouldn't since the warm up requests overlapped with the actual requests
which pushed requests past allowed limits by the app. We now wait a
second after the warm up requests before starting the actual test.

We also needed to upgrade to the latest dropwizard 4.0 version.