Update build image docker dependencies to latest supported version
All 3 docker dependencies are on the latest version we support. Sysbox
has an issue that doesn't work beyond this containerd.io version. The
java docker lib doesn't work beyond the docker-ce version.

Additionally to note, ubuntu 24.04 doesn't go down to a version this low
for docker-ce, so we can't upgrade to that.